Fix: inconsistent edge thickness of communication channels
When modeling communication channels and zooming to, say, 200%,
the edges of communication channels have different thickness.

Apparently the actual stroke width used to draw the first edge
differs from the getCurrentStrokeWidth() call issued before
drawing the channel which is then used when drawing the second
edge.

This is a quick change addressing the issue.